# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ipr": {
            "post": {
                "description": "Returns the IPR(s) for the user. If the date parameter is not passed into the body or is invalid, the most recent IPR is returned.\nIt is important the format of the date follows the format \"01/02/2006\" (01 = month, 02 = day, 2006 = year), with leading zeros like shown in the format.\nFor all possible dates, refer to the \"/ipr/all\" endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipr"
                ],
                "parameters": [
                    {
                        "description": "Body Params",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.IprRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IPRResponse"
                        }
                    }
                }
            }
        },
        "/ipr/all": {
            "post": {
                "description": "Returns all the IPRs for the user, or just the dates depending on the DatesOnly parameter's value in the body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ipr"
                ],
                "parameters": [
                    {
                        "description": "Body Params",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.IprAllRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IPRResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Class": {
            "type": "object",
            "properties": {
                "course": {
                    "description": "The course ID of the class",
                    "type": "string"
                },
                "name": {
                    "description": "The name of the class",
                    "type": "string"
                },
                "period": {
                    "description": "What period the class is for the student, relative to the current schedule",
                    "type": "string"
                },
                "room": {
                    "description": "The room number of the class",
                    "type": "string"
                },
                "teacher": {
                    "description": "The name of the teacher of the class",
                    "type": "string"
                }
            }
        },
...
        "models.IPR": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "The date the IPR was submitted",
                    "type": "string"
                },
                "entries": {
                    "description": "An array representing all the IPR entries",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IPREntry"
                    }
                }
            }
        },
        "models.IPREntry": {
            "type": "object",
            "properties": {
                "class": {
                    "description": "Information about the class related to the IPREntry",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Class"
                        }
                    ]
                },
                "grade": {
                    "description": "The average at the moment the progress report was submitted",
                    "type": "string"
                }
            }
        },
        "models.IPRResponse": {
            "type": "object",
            "properties": {
                "err": {
                    "description": "If there was an error",
                    "type": "boolean"
                },
                "ipr": {
                    "description": "The resulting IPR(s)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IPR"
                    }
                },
                "msg": {
                    "description": "The associated message",
                    "type": "string"
                }
            }
        },
        "models.IprAllRequestBody": {
            "type": "object",
            "required": [
                "base",
                "password",
                "username"
            ],
            "properties": {
                "base": {
                    "description": "The base URL for the PowerSchool HAC service",
                    "type": "string",
                    "minLength": 1,
                    "example": "homeaccess.katyisd.org"
                },
                "datesOnly": {
                    "description": "Whether to return only dates or all the IPRs",
                    "type": "boolean",
                    "default": false,
                    "example": true
                },
                "password": {
                    "description": "The password to log in with",
                    "type": "string",
                    "minLength": 1,
                    "example": "j382704"
                },
                "username": {
                    "description": "The username to log in with",
                    "type": "string",
                    "minLength": 1,
                    "example": "j1732901"
                }
            }
        },
        "models.IprRequestBody": {
            "type": "object",
            "required": [
                "base",
                "password",
                "username"
            ],
            "properties": {
                "base": {
                    "description": "The base URL for the PowerSchool HAC service",
                    "type": "string",
                    "minLength": 1,
                    "example": "homeaccess.katyisd.org"
                },
                "date": {
                    "description": "The date of the IPR to return",
                    "type": "string",
                    "example": "09/06/2022"
                },
                "password": {
                    "description": "The password to log in with",
                    "type": "string",
                    "minLength": 1,
                    "example": "j382704"
                },
                "username": {
                    "description": "The username to log in with",
                    "type": "string",
                    "minLength": 1,
                    "example": "j1732901"
                }
            }
        },
...
    }
}